+++ daemon/config/config.go
@@ -72,6 +72,22 @@ const (
 	LibnetDataPath = "network/files"
 )
 
+// Features get serialized into the `Engine-Features` header by the `/_ping` handler,
+// so we need some limits to ensure that we can serialize them into the header.
+const (
+	// The HTTP spec does not define size limits for headers, but clients usually set limits.
+	// The Go default HTTP client has a default of 1MB:
+	// https://cs.opensource.google/go/go/+/refs/tags/go1.23.3:src/net/http/server.go;l=916
+	// but other clients and proxies usually have lower limits. To be safe, we should try to
+	// set a conservative limit.
+	// 100*(30 +1 (=) +1 (,) +5 (false/true)) = 3700 which is under 4K.
+
+	// MaxFeatures is the maximum number of configured daemon features.
+	MaxFeatures = 100
+	// MaxFeatureKeyLen is the maximum length for feature names.
+	MaxFeatureKeyLen = 30
+)
+
 // flatOptions contains configuration keys
 // that MUST NOT be parsed as deep structures.
 // Use this to differentiate these options
@@ -665,6 +681,24 @@ func ValidateMinAPIVersion(ver string) error {
 	return nil
 }
 
+func validateFeatures(features map[string]bool) error {
+	if len(features) > MaxFeatures {
+		return errors.Errorf("too many features – expected max %d, found %d", MaxFeatures, len(features))
+	}
+	for k := range features {
+		if strings.Contains(k, "=") {
+			return errors.Errorf("invalid feature – key cannot contain '=': %s", k)
+		}
+		if strings.Contains(k, ",") {
+			return errors.Errorf("invalid feature – key cannot contain ',': %s", k)
+		}
+		if len(k) > MaxFeatureKeyLen {
+			return errors.Errorf("feature name length cannot be over %d: %s", MaxFeatureKeyLen, k)
+		}
+	}
+	return nil
+}
+
 // Validate validates some specific configs.
 // such as config.DNS, config.Labels, config.DNSSearch,
 // as well as config.MaxConcurrentDownloads, config.MaxConcurrentUploads and config.MaxDownloadAttempts.
@@ -730,6 +764,12 @@ func Validate(config *Config) error {
 
 	for _, h := range config.Hosts {
 		if _, err := opts.ValidateHost(h); err != nil {
+			return err
+		}
+	}
+
+	if len(config.Features) > 0 {
+		if err := validateFeatures(config.Features); err != nil {
 			return err
 		}
 	}